Added config.py as a config file for database logins. Added cheatsheet datatypes.py. This is not complete. Small tweaks in functions.py for minor bugs.
Todo:
Complete remaining functions.
Improve comments.
Keep testing functions.

Plans:
Want to it so when I insert into a table, it checks if a column is missing, and if it does, it will be added (I may make this function separately)
Need to find a way of inserting written text properly. Currently can't use words like I've or you're etc.